Agora apenas os proprios usuarios podem deletar, atualizar e ver seu proprio perfil e endereço

+++ Services/Address/DeleteAddressService.cs
@@ -1,5 +1,6 @@
 ﻿using Api.KmgShop.UserManager.Models;
 using Api.KmgShop.UserManager.Repository;
+using System.Security.Claims;
 
 namespace Api.KmgShop.UserManager.Services.DeleteAddress;
 
@@ -12,10 +13,13 @@ public class DeleteAddressService
         _addressRepository = addressRepository;
     }
 
-    public async Task<Address> DeleteAddressAsync(int addressId)
+    public async Task<Address> DeleteAddressAsync(int addressId, ClaimsPrincipal userClaims)
     {
         var address = await _addressRepository.GetAddressByIdAsync(addressId);
         if (address == null) return null;
+
+        int idUser = int.Parse(userClaims.FindFirstValue("id"));
+        if (idUser != address.UserId) return null;
 
         await _addressRepository.DeleteAddressAsync(address);
         return address;
