+++ Services/User/DeleteUserService.cs
@@ -1,5 +1,6 @@
 ﻿using Api.KmgShop.UserManager.Models;
 using Api.KmgShop.UserManager.Repository;
+using System.Security.Claims;
 
 namespace Api.KmgShop.UserManager.Services.DeleteUser;
 
@@ -12,10 +13,13 @@ public class DeleteUserService
         _userRepository = userRepository;
     }
 
-    public async Task<User> DeleteUserAsync(int userId)
+    public async Task<User> DeleteUserAsync(int userId, ClaimsPrincipal userClaims)
     {
         var user = await _userRepository.GetUserByIdAsync(userId);
         if(user == null) return null;
+
+        int idUser = int.Parse(userClaims.FindFirstValue("id"));
+        if (idUser != user.UserId) return null;
 
         await _userRepository.DeleteUserAsync(user);
         return user;
